Split email modal into separate component
